+++ Psorcast/Psorcast/TreatmentSelectionStepViewController.swift
@@ -35,74 +35,6 @@ import Foundation
 import BridgeApp
 import BridgeAppUI
 
-public class TreatmentSelectionStepObject: RSDUIStepObject, RSDStepViewControllerVendor {
-        
-    private enum CodingKeys: String, CodingKey, CaseIterable {
-        case items
-    }
-    
-    public var items = [TreatmentItem]() {
-        didSet {
-            self.refreshSorting()
-        }
-    }
-    
-    public private (set) var sortedItems = [String: [TreatmentItem]]()
-    public private (set) var sortedSections = [String]()
-    
-    open var otherSectionIdentifier: String {
-        return Localization.localizedString("OTHER_TREATMENT_SECTION_TITLE")
-    }
-    
-    private func refreshSorting() {
-        let otherSectionId = self.otherSectionIdentifier
-        // Sort the items grouped by sectionIdentifier
-        // First, get the sorted unique set of sectionIdentifiers
-        let sectionArray = self.items.map({ $0.sectionIdentifier ?? otherSectionId })
-        let sectionSet = Set(sectionArray).sorted(by: { $0 < $1 })
-        self.sortedItems = [String: [TreatmentItem]]()
-        self.sortedSections = Array(sectionSet)
-        // Then, build the sorted map
-        for sectionIdentifier in sectionSet {
-            self.sortedItems[sectionIdentifier] = self.items.filter({ ($0.sectionIdentifier ?? otherSectionId) == sectionIdentifier })
-        }
-    }
-    
-    /// Default type is `.treatmentSelection`.
-    open override class func defaultType() -> RSDStepType {
-        return .treatmentSelection
-    }
-    
-    public required init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        self.items = try container.decode([TreatmentItem].self, forKey: .items)
-        
-        try super.init(from: decoder)
-        
-        self.refreshSorting()
-    }
-    
-    required public init(identifier: String, type: RSDStepType? = nil) {
-        super.init(identifier: identifier, type: type)
-        self.self.refreshSorting()
-    }
-    
-    /// Override to set the properties of the subclass.
-    override open func copyInto(_ copy: RSDUIStepObject) {
-        super.copyInto(copy)
-        guard let subclassCopy = copy as? TreatmentSelectionStepObject else {
-            assertionFailure("Superclass implementation of the `copy(with:)` protocol should return an instance of this class.")
-            return
-        }
-        subclassCopy.items = self.items
-    }
-    
-    public func instantiateViewController(with parent: RSDPathComponent?) -> (UIViewController & RSDStepController)? {
-        return TreatmentSelectionStepViewController(step: self, parent: parent)
-    }
-}
-
 public class TreatmentSelectionStepViewController: RSDStepViewController, UITableViewDelegate, UITableViewDataSource, UISearchBarDelegate {
     
     @IBOutlet weak var tableView: UITableView!
@@ -514,194 +446,3 @@ public class TreatmentSelectionStepViewController: RSDStepViewController, UITabl
         super.goForward()
     }
 }
-
-public struct TreatmentItem: Codable {
-    public var identifier: String
-    public var detail: String?
-    public var sectionIdentifier: String?
-}
-
-open class TreatmentSelectionTableViewCell: RSDSelectionTableViewCell {
-    
-    @IBOutlet public var removeImage: UIImageView?
-    
-    public override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
-        // Add the line separator
-        removeImage = UIImageView()
-        removeImage?.image = UIImage(named: "ClearIcon")
-        contentView.addSubview(removeImage!)
-        
-        removeImage!.translatesAutoresizingMaskIntoConstraints = false
-        removeImage!.rsd_makeWidth(.equal, 18.0)
-        removeImage!.rsd_makeHeight(.equal, 18.0)
-        removeImage!.rsd_alignToSuperview([.trailing], padding: 24.0)
-        removeImage!.rsd_alignCenterVertical(padding: 0.0)
-    }
-    
-    required public init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-    }
-}
-
-open class TreatmentSelectionTableHeader: UITableViewHeaderFooterView, RSDViewDesignable {
-        
-    internal let kHeaderHorizontalMargin: CGFloat = 28.0
-    internal let kHeaderVerticalMargin: CGFloat = 8.0
-    
-    public var backgroundColorTile: RSDColorTile?
-    
-    public var designSystem: RSDDesignSystem?
-    
-    @IBOutlet public var titleLabel: UILabel?
-    @IBOutlet public var separatorLine: UIView?
-
-    open private(set) var titleTextType: RSDDesignSystem.TextType = .mediumHeader
-    
-    public func setDesignSystem(_ designSystem: RSDDesignSystem, with background: RSDColorTile) {
-        updateColorsAndFonts()
-    }
-    
-    func updateColorsAndFonts() {
-        let designSystem = self.designSystem ?? RSDDesignSystem()
-        let background = self.backgroundColorTile ?? RSDGrayScale().white
-        
-        contentView.backgroundColor = background.color
-        separatorLine?.backgroundColor = designSystem.colorRules.separatorLine
-        titleLabel?.textColor = designSystem.colorRules.textColor(on: background, for: titleTextType)
-        titleLabel?.font = designSystem.fontRules.font(for: titleTextType, compatibleWith: traitCollection)
-    }
-    
-    public override init(reuseIdentifier: String?) {
-        super.init(reuseIdentifier: reuseIdentifier)
-        
-        self.backgroundColor = UIColor.white
-        
-        // Add the title label
-        titleLabel = UILabel()
-        contentView.addSubview(titleLabel!)
-        
-        titleLabel!.translatesAutoresizingMaskIntoConstraints = false
-        titleLabel!.numberOfLines = 0
-        titleLabel!.textAlignment = .left
-
-        titleLabel!.rsd_alignToSuperview([.leading], padding: kHeaderHorizontalMargin)
-        titleLabel!.rsd_align([.trailing], .lessThanOrEqual, to: contentView, [.trailing], padding: kHeaderHorizontalMargin, priority: .required)
-        titleLabel!.rsd_alignToSuperview([.top], padding: kHeaderVerticalMargin, priority: UILayoutPriority(rawValue: 700))
-        titleLabel!.rsd_alignToSuperview([.bottom], padding: kHeaderVerticalMargin)
-        
-        // Add the line separator
-        separatorLine = UIView()
-        separatorLine!.backgroundColor = UIColor.lightGray
-        contentView.addSubview(separatorLine!)
-        
-        separatorLine!.translatesAutoresizingMaskIntoConstraints = false
-        separatorLine!.rsd_alignToSuperview([.leading, .bottom, .trailing], padding: 0.0)
-        separatorLine?.rsd_makeHeight(.equal, 1.0)
-        
-        updateColorsAndFonts()
-        setNeedsUpdateConstraints()
-    }
-    
-    public required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        updateColorsAndFonts()
-    }
-}
-
-extension UITextField {
-    func setLeftPaddingPoints(_ amount : CGFloat) {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
-        self.leftView = paddingView
-        self.leftViewMode = .always
-    }
-    
-    func setRightPaddingPoints(_ amount : CGFloat) {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: amount, height: self.frame.size.height))
-        self.rightView = paddingView
-        self.rightViewMode = .always
-    }
-}
-
-class LevensteinTools {
-    // return minimum value in a list of Ints
-    fileprivate class func minNum(_ numbers: Int...) -> Int {
-        return numbers.reduce(numbers[0], {$0 < $1 ? $0 : $1})
-    }
-
-    class func levenshtein(aStr: String, bStr: String) -> Int {
-        // create character arrays
-        let a = Array(aStr)
-        let b = Array(bStr)
-
-        // initialize matrix of size |a|+1 * |b|+1 to zero
-        var dist = [[Int]]()
-        for _ in 0...a.count {
-            dist.append([Int](repeating: 0, count: b.count + 1))
-        }
-
-        // 'a' prefixes can be transformed into empty string by deleting every char
-        for i in 1...a.count {
-            dist[i][0] = i
-        }
-
-        // 'b' prefixes can be created from empty string by inserting every char
-        for j in 1...b.count {
-            dist[0][j] = j
-        }
-
-        for i in 1...a.count {
-            for j in 1...b.count {
-                if a[i-1] == b[j-1] {
-                    dist[i][j] = dist[i-1][j-1]  // noop
-                } else {
-                    dist[i][j] = LevensteinTools.minNum(
-                        dist[i-1][j] + 1,  // deletion
-                        dist[i][j-1] + 1,  // insertion
-                        dist[i-1][j-1] + 1  // substitution
-                    )
-                }
-            }
-        }
-
-        return dist[a.count][b.count]
-    }
-    
-    ///
-    /// This function uses various phrases created from title and detail to calculate
-    /// the best, or min, levenstein score for the parameters.
-    /// This function favors favor prefixes, because the user will start typing word from the first letter
-    ///
-    class func minLevensteinScore(for searchText: String, titleText: String, detailText: String?) -> Int {
-        
-        if searchText.count == 0 {
-            return Int.max
-        }
-        var minScore = Int.max
-
-        var fullWordArray = [String]()
-        // Build the list of phrases we will test levenstein score against threshold.
-        for text in [titleText, detailText ?? ""] {
-            // Ignore empty strings
-            if text.count > 0 {
-                fullWordArray.append(text)
-                let splitWords = text.split(separator: " ")
-                // Ignore the first word, but add all other words after that.
-                if splitWords.count > 1 {
-                    fullWordArray.append(contentsOf: Array(splitWords.map({ String($0) })))
-                }
-            }
-        }
-                
-        for text in fullWordArray {
-            // Truncate the phrase to meet length of search text to also favor prefixes.
-            // Also, make everything lowercase to ignore capitalization differences.
-            let treatmentText = String(text.lowercased().prefix(searchText.count))
-            minScore = min(minScore, LevensteinTools.levenshtein(aStr: searchText, bStr: treatmentText))
-        }
-        
-        return minScore
-    }
-}
-
